server/requirements.txt
Flask>=2.0
python-dotenv>=0.21
requests>=2.28
